+++ lib/src/extensions/string_extensions.dart
@@ -5,27 +5,31 @@ import 'package:flutter/services.dart';
 
 extension TraceXStringExt on String {
   Future<void> copyToClipboard(BuildContext context) {
+    final snackBar = SnackBar(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(8.0),
+      ),
+      duration: const Duration(seconds: 1),
+      behavior: SnackBarBehavior.floating,
+      dismissDirection: DismissDirection.startToEnd,
+      content: Text(
+        'Copied',
+        style: Theme.of(context)
+            .textTheme
+            .bodyMedium
+            ?.copyWith(color: Colors.white),
+        textAlign: TextAlign.center,
+      ),
+      margin: EdgeInsets.only(
+        bottom: MediaQuery.of(context).size.height - 160,
+        right: 20,
+        left: 20,
+      ),
+    );
+
     ScaffoldMessenger.of(context)
       ..hideCurrentSnackBar()
-      ..showSnackBar(
-        SnackBar(
-          elevation: 0.0,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(8.0),
-          ),
-          duration: const Duration(seconds: 1),
-          behavior: SnackBarBehavior.floating,
-          content: Center(
-            child: Text(
-              'Copied',
-              style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                color: Colors.white,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ),
-        ),
-      );
+      ..showSnackBar(snackBar);
 
     HapticFeedback.lightImpact();