+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 1.0.0
 
 Initial release
+
+## 1.1.0
+
+Implement json highlight
+
+## 1.1.1
+
+Implement json highlight
+Update readme